requirements.txt
Flask==2.3.2
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
nltk==3.8.1
PyPDF2==3.0.1
python-docx==0.8.11
werkzeug==2.3.6
matplotlib==3.7.1
seaborn==0.12.2
